% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rr.R
\name{calculate_rr_in_df}
\alias{calculate_rr_in_df}
\title{Calculate Relative Risk (RR) from a Data Frame}
\usage{
calculate_rr_in_df(
  df,
  column_cases_exposed = "cases_exposed",
  column_total_exposed = "total_exposed",
  column_cases_unexposed = "cases_unexposed",
  column_total_unexposed = "total_unexposed"
)
}
\arguments{
\item{df}{A data frame containing the following columns:
\itemize{
\item \code{cases_exposed}: Number of cases in the exposed group.
\item \code{total_exposed}: Total number of individuals in the exposed group.
\item \code{cases_unexposed}: Number of cases in the unexposed group.
\item \code{total_unexposed}: Total number of individuals in the unexposed group.
}}

\item{column_cases_exposed}{Name of the column containing the number of cases in the exposed group. Default is "cases_exposed".}

\item{column_total_exposed}{Name of the column containing the total number of individuals in the exposed group. Default is "total_exposed".}

\item{column_cases_unexposed}{Name of the column containing the number of cases in the unexposed group. Default is "cases_unexposed".}

\item{column_total_unexposed}{Name of the column containing the total number of individuals in the unexposed group. Default is "total_unexposed".}
}
\value{
A data frame with an additional column for relative risk (RR).
}
\description{
This function calculates the relative risk (RR) from a data frame containing counts of cases and total individuals in both exposed and unexposed groups.
}
