% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_toy_dataset.R
\docType{data}
\name{toy_dataset}
\alias{toy_dataset}
\title{toy_dataset}
\format{
A data frame with 4 rows and 4 variables:
\describe{
\item{ cases_exposed }{  integer }
\item{ total_exposed }{  integer }
\item{ cases_unexposed }{  integer }
\item{ total_unexposed }{  integer }
}
}
\source{
Source
}
\usage{
toy_dataset
}
\description{
Description.
}
\keyword{datasets}
